FEniCS backend: Move parameters code
