+++ docs/source/conf.py
@@ -12,6 +12,7 @@ autoapi_ignore = ["*/checkpoint_schedules/__init__.py",
                   "*/fenics/backend.py",
                   "*/fenics/backend_code_generator_interface.py",
                   "*/fenics/backend_patches.py",
+                  "*/fenics/parameters.py",
                   "*/firedrake/__init__.py",
                   "*/firedrake/backend.py",
                   "*/firedrake/backend_code_generator_interface.py",
