+++ tlm_adjoint/fenics/__init__.py
@@ -12,6 +12,11 @@ from .caches import *              # noqa: E402,F401
 from .equations import *           # noqa: E402,F401
 from .functions import *           # noqa: E402,F401
 from .interpolation import *       # noqa: E402,F401
+from .parameters import *          # noqa: E402,F401
 from .projection import *          # noqa: E402,F401
 
 from .backend_patches import *     # noqa: E402,F401
+
+del backend, backend_code_generator_interface, backend_interface, caches, \
+    equations, functions, interpolation, parameters, projection, \
+    backend_patches  # noqa: F821
